% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffogram.R
\name{diffogram}
\alias{diffogram}
\title{A diffogram for lsmeans}
\usage{
diffogram(lsmodel, logt = FALSE)
}
\arguments{
\item{lsmodel}{an object from lsmeans function}

\item{logt}{are the reposnce after log transformation?}
}
\description{
This function plots a diffogram for least square means calculated
with the lsmeans package.
The diffogram presents results from post hoc testing.
More informations about diffograms:
,,Interpreting the Differences Among LSMEANS in Generalized Linear Models''
http://www.mwsug.org/proceedings/2011/dataviz/MWSUG-2011-DG08.pdf
}
\examples{
\dontrun{
model <- lmer(log(length) ~ group + (1|Animal) , data=data)
lsmodel <- lsmeans(model,  pairwise~group, adjust="tukey")
diffogram(lsmodel)
}
}

