% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_raport.R
\name{generate_raport}
\alias{generate_raport}
\title{A generation of a raport}
\usage{
generate_raport(folder_path, file_path, animal_col_name, group_col_name,
   photo_col_name = "Photo_ID_rel", spines_col_name,
   properties_col_name = "length", x_lim = c(0, 2),
   strat = paste0(animal_col_name, ":group"), mixed = TRUE, ...)
}
\arguments{
\item{folder_path}{a path to the folder where raport will be made}

\item{file_path}{a path to the file with spines data}

\item{animal_col_name}{a name of column with animal}

\item{group_col_name}{a name of column with group}

\item{photo_col_name}{a name of column with photos; default: Photo_ID_rel}

\item{spines_col_name}{a name of column with spines' numbers}

\item{properties_col_name}{a name/names of column/s with properties of
dendritic spine which will be analysed; default "length"}

\item{x_lim}{a vector with limits of x axis; default: c(0, 2)}

\item{strat}{a stratification (Animal); default: paste0(animal_col_name, ":group")}

\item{mixed}{if TRUE a mixed model is applied, if FALSE it's just standard
linear regression; default: TRUE}

\item{...}{other arguments like sep, header, sheet, etc.}
}
\value{
invisible NULL
}
\description{
Function \code{generate_raport} generates a raport.
}
\details{
Function \code{generate_raport} generates
a raport in indicated folder. There will be made two files:
*.Rmd and *.html. The scheme of the names of the files is:
"spines-raport-YYYY-MM-DD.(Rmd|html)", where:
\itemize{
\item "YYYY" is year,
\item "MM" is month,
\item "DD" is day.}
The title of the raport is setted to "Spines". If you want change
its title or add author to the raport edit the raport file (*.Rmd) by hand.
The space with these informations is at the top of the file.
When the raport file is ready, use \code{render_raport} function
to render it again to the *.html file.
}

