% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_crossed_effects.R
\name{model_crossed_effects}
\alias{model_crossed_effects}
\title{Function model crossed effects and returns two data frames with model summaries}
\usage{
model_crossed_effects(data, var, trans = I, f1 = "group",
  f2 = "condition", strat = "Animal", mixed = TRUE)
}
\arguments{
\item{data}{data.frame with data}

\item{var}{variable of interest}

\item{trans}{transformation of the var variable before ANOVA}

\item{f1}{first effect}

\item{f2}{second effect}

\item{strat}{stratification (Animal)}

\item{mixed}{if TRUE a mixed model is applied, if FALSE it's just standard linear regression}
}
\description{
Function calculates models and computes post hoc tests for these models
It is assumed that there are two crossed effects.
}

