% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_animals.R
\name{plot_animals}
\alias{plot_animals}
\title{Plot of distributions of spines}
\usage{
plot_animals(spines, property = "length", box = FALSE)
}
\arguments{
\item{spines}{a data.frame of spines class}

\item{property}{a character with property variable; default: "length"}

\item{box}{a way of presenting distribution - FALSE for kernel density
or TRUE for boxplot; default: FALSE}
}
\value{
a panel plot of distributions
}
\description{
Function \code{plot_animals} plots densities for given variable
divided into groups plotted on different panels for different animal
}

