% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distribution.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Distribution of spines}
\usage{
plot_ecdf(spines, properties = "length", ecdf = TRUE)
}
\arguments{
\item{spines}{a data.frame of spines class}

\item{properties}{a vector with properties variable/s}

\item{ecdf}{if TRUE then cumulative density function is plotted,
if FALSE then density function is plotted}
}
\value{
a (cumulative) density function plot
}
\description{
Function \code{plot_ecdf} plots a distribution (both density function
and cumulative density function) for dendritic spines grouped by one
or more grouping variables.
}

