% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render_raport.R
\name{render_raport}
\alias{render_raport}
\title{A render of raport}
\usage{
render_raport(file_path, destination_path)
}
\arguments{
\item{file_path}{a path to the file with generation of raport}

\item{destination_path}{a path to the file with raport}
}
\value{
invisible NULL
}
\description{
Function \code{render_raport} renders
a raport.
}

