% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{A summary of loaded data}
\usage{
summary(spines)
}
\arguments{
\item{spines}{a data.frame of spines class}
}
\value{
invisible NULL
}
\description{
Function \code{summary} summarises loaded data
by \code{read_spines} function.
}

