% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.spines}
\alias{summary.spines}
\title{A summary of loaded data}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{a data.frame of spines class}

\item{...}{other arguments}
}
\value{
invisible NULL
}
\description{
Function \code{summary} summarises loaded data
by \code{read_spines} function.
}

