% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformations.R
\name{transformations}
\alias{transformations}
\title{A choose of transformations for property}
\usage{
transformations(property, as_character = FALSE)
}
\arguments{
\item{property}{a character with property variable}

\item{as_character}{a logical value indicating whether value on return
must be a character vector; default: FALSE}
}
\value{
a list with two functions or character vector
}
\description{
Function \code{transformations} chooses
a tranformations for chosen property.
}
\details{
Function \code{transformations} chooses
a tranformations for chosen property. The first tranformation
is a function for applying the property to the model.
The second is a inversion of first function.
}

