% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffogram.R
\name{diffogram}
\alias{diffogram}
\title{A diffogram for lsmeans}
\usage{
diffogram(lsmodel)
}
\arguments{
\item{lsmodel}{an object from lsmeans function}
}
\description{
This function plots a diffogram for least square means calculated
with the lsmeans package.
The diffogram presents results from post hoc test.
}
\examples{
\dontrun {
model <- lmer(log(length) ~ group + (1|Animal) , data=data)
lsmodel <- lsmeans(model,  pairwise~group, adjust="tukey")
diffogram(lsmodel)
}
}

