% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_raport_file.R
\name{generate_raport_file}
\alias{generate_raport_file}
\title{A generation of raport file}
\usage{
generate_raport_file(folder_path, spines, x_lim = c(0, 2))
}
\arguments{
\item{folder_path}{a path to the folder where file will be made}

\item{file_path}{a path to the file with spines data}

\item{spines}{a data.frame of spines class}

\item{x_lim}{a vector with limits of x axis; default: c(0, 2)}
}
\value{
invisible NULL
}
\description{
Function \code{generate_raport_file} generates
a raport file.
}

