% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_crossed_effects.R
\name{plot_crossed_effects}
\alias{plot_crossed_effects}
\title{Function plots crossed effects}
\usage{
plot_crossed_effects(data, var, trans = I, inv = I,
   strat = "Animal", mixed = TRUE, addpoints = FALSE)
}
\arguments{
\item{spines}{a data.frame of spines class}

\item{property}{a variable of interest}

\item{trans}{a transformation of the property variable before ANOVA; default: I}

\item{inv}{an inverse transformation of the trans(property); default: I}

\item{strat}{a stratification (Animal); default: "Animal"}

\item{mixed}{if TRUE a mixed model is applied, if FALSE it's just standard
linear regression; default: TRUE}

\item{addpoints}{if TRUE additional points for each group/condition/strat group
will be added; default: FALSE}
}
\value{
a crossed effects plot
}
\description{
Function \code{plot_crossed_effects} calculates averages and sd's
(with the use of fixed or mixed models) and plots them.
It is assumed that there are two crossed effects.
}

