% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Function plots crossed effects}
\usage{
plot_distribution(data, var, f1 = "Group", strat = "Animal", box = FALSE)
}
\arguments{
\item{data}{data.frame with data}

\item{var}{variable of interest}

\item{f1}{first effect (used as fill)}

\item{strat}{stratification (Animal)}

\item{box}{how to present distribution? boxplot or kernel density?}
}
\description{
Function plots densities for given varialbe divided into groups (f1) plotted on different panels (strat)
}

