% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot of distributions of spines}
\usage{
plot_distributions(spines, property = "length", box = FALSE)
}
\arguments{
\item{property}{a variable of interest; default: "length"}

\item{box}{a way of presenting distribution - FALSE for kernel density
or TRUE for boxplot; default: FALSE}

\item{data}{a data.frame of spines class}
}
\value{
a panel plot of distributions
}
\description{
Function \code{plot_distributions} plots densities for given variable
divided into groups plotted on different panels for different animal
}

