% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Distribution of spines}
\usage{
plot_distributions(spines, property = "length", ecdf = TRUE, x_lim = c(0, 2))
}
\arguments{
\item{spines}{a data.frame of spines class}

\item{property}{a character with property variable; default: "length"}

\item{ecdf}{if TRUE then cumulative density function is plotted,
if FALSE then density function is plotted; default: TRUE}

\item{x_lim}{a vector with limits of x axis; default: c(0, 2)}
}
\value{
a (cumulative) density function plot
}
\description{
Function \code{plot_distributions} plots a distribution (both density function
and cumulative density function) for dendritic spines grouped by one
or more grouping variables.
}

