% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distribution.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Distribution of spines}
\usage{
plot_ecdf(spines, group, ecdf = TRUE)
}
\arguments{
\item{spines}{a vector with spine properties (length, area etc)}

\item{group}{a vector with grouping variable}

\item{ecdf}{if TRUE then cumulative density function is plotted, if FALSE then density function is plotted}
}
\description{
This function plots a distribution (both density function and cumulative density function)
for dendritic spines grouped by one or more grouping variables.
}

