% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_spines.R
\name{read_spines}
\alias{read_spines}
\title{Proper loading data}
\usage{
read_spines(file, animal_col_name, group_col_name,
   photo_col_name = "Photo_ID_rel", spines_col_name,
   properties_col_name = "length", ...)
}
\arguments{
\item{file}{a file's path that will be loaded}

\item{animal_col_name}{a name of column with animal}

\item{group_col_name}{a name of column with group}

\item{photo_col_name}{a name of column with photos; default: Photo_ID_rel}

\item{spines_col_name}{a name of column with spines' numbers}

\item{properties_col_name}{a name/names of column/s with properties of
dendritic spine which will be analysed; default "length"}

\item{...}{other arguments like sep, header, sheet, etc.}
}
\value{
a data.frame of spines class
}
\description{
Function \code{read_spines} loads data from the indicated file.
}
\details{
Function \code{read_spines} loads data from the indicated file.
The result of this function is a data frame with filtred data
including chosen by user columns and two new columns - 'group'
and 'condition'. In 'group' column is a group of the animal
and in 'condition' is substance which was used in the study.
If there was only group (pattern like 'ab', not like 'a(separator)b')
in group_col_name column, a 'condition' is setted on 'x',
so remember to update data file before using this function.
Also remember to do not use 'group' as a name for
group_col_name column, because 'group' column
will be added by \code{read_spines} function and it
will be used by plotting functions.
}

