% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render_raport.R
\name{render_raport}
\alias{render_raport}
\title{A render of raport}
\usage{
render_raport(file_path, destination_path)
}
\arguments{
\item{file_path}{a path to the file with generation of raport}

\item{destination_path}{a path to the file with raport}
}
\value{
invisible NULL
}
\description{
Function \code{render_raport} renders
a raport.
}
\details{
Function \code{render_raport} renders
a raport in indicated folder. To generate raport the *.Rmd
file with generation of raport is needed. The scheme
of this file is: "spines-raport-YYYY-MM-DD.Rmd", where:
\itemize{
\item "YYYY" is year,
\item "MM" is month,
\item "DD" is day.}
It is possible to change raport's title or add author to the raport.
To do that, edit the raport file by hand. The space with these
informations is at the top of the file. When the raport file is ready,
use \code{render_raport} function to render it to the *.html file.
}

