% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.spines}
\alias{summary.spines}
\title{A summary of loaded data}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{a data.frame of spines class}

\item{...}{other arguments}
}
\value{
invisible NULL
}
\description{
Function \code{summary} summarises loaded data
by \code{read_spines} function.
}
\details{
Function \code{summary} summarises loaded data
by \code{read_spines} function.

Summarised columns:
\enumerate{
\item Animal column - (unique) appearances of animals,
\item Group column - appearance of elements,
\item group and condition columns - appearance of elements (unique appearance),
\item properties columns - statistic summary.
}
}

